+++ resources/views/admin/clients/create.blade.php
@@ -6,8 +6,8 @@ New Client « Admin CP
 
 @section('content')
 <h1>New Client</h1>
-<form action="/admin/clients/new" method="post" accept-charset="utf-8">
-  <input type="hidden" name="_token" value="{{ csrftoken() }}">
+<form action="/admin/clients/" method="post" accept-charset="utf-8">
+  {{ csrf_field() }}
   <input type="text" name="client_url" id="client_url" placeholder="client_url"><br>
   <input type="text" name="client_name" id="client_name" placeholder="client_name"><br>
   <input type="submit" name="submit" value="Submit">