+++ resources/views/admin/clients/editsuccess.blade.php
@@ -1,9 +1,0 @@
-@extends('master')
-
-@section('title')
-Edit Client « Admin CP
-@stop
-
-@section('content')
-<p>Successfully edited the client information.</p>
-@stop