+++ resources/views/admin/contacts/editsuccess.blade.php
@@ -1,9 +1,0 @@
-@extends('master')
-
-@section('title')
-Edit Contact « Admin CP
-@stop
-
-@section('content')
-<p>Successfully edited the contact information.</p>
-@stop