+++ resources/views/admin/contacts/newsuccess.blade.php
@@ -1,9 +1,0 @@
-@extends('master')
-
-@section('title')
-New Contact « Admin CP
-@stop
-
-@section('content')
-<p>Succesfully created new contact entry.</p>
-@stop