+++ resources/views/admin/notes/newsuccess.blade.php
@@ -1,9 +1,0 @@
-@extends('master')
-
-@section('title')
-New Note Success « Admin CP
-@stop
-
-@section('content')
-<p>Successfully created note with id: {{ $id }}. {{ $shorturl }}</p>
-@stop