+++ resources/views/admin/places/editsuccess.blade.php
@@ -1,9 +1,0 @@
-@extends('master')
-
-@section('title')
-Edit Place « Admin CP
-@stop
-
-@section('content')
-<p>Successfully edited the place information.</p>
-@stop