+++ resources/views/admin/places/newsuccess.blade.php
@@ -1,9 +1,0 @@
-@extends('master')
-
-@section('title')
-New Place « Admin CP
-@stop
-
-@section('content')
-<p>Succesfully created new place info.</p>
-@stop