+++ resources/views/admin/welcome.blade.php
@@ -8,17 +8,17 @@ Admin CP
 <h1>Hello {{ $name }}!</h1>
 
 <h2>Articles</h2>
-<p>You can either <a href="/admin/articles/new">create</a> new blog posts, or <a href="/admin/articles/edit">edit</a> them.<p>
+<p>You can either <a href="/admin/articles/create">create</a> new blog posts, or <a href="/admin/articles/">edit</a> them.<p>
 
 <h2>Notes</h2>
-<p>You can either <a href="/admin/notes/new">create</a> new notes, or <a href="/admin/notes/edit">edit</a> them.<p>
+<p>You can either <a href="/admin/notes/create">create</a> new notes, or <a href="/admin/notes/">edit</a> them.<p>
 
 <h2>Clients</h2>
-<p>You can either <a href="/admin/clients/new">create</a> new contacts, or <a href="/admin/contacts/edit">edit</a> them.</p>
+<p>You can either <a href="/admin/clients/create">create</a> new contacts, or <a href="/admin/clients/">edit</a> them.</p>
 
 <h2>Contacts</h2>
-<p>You can either <a href="/admin/contacts/new">create</a> new contacts, or <a href="/admin/contacts/edit">edit</a> them.</p>
+<p>You can either <a href="/admin/contacts/create">create</a> new contacts, or <a href="/admin/contacts/">edit</a> them.</p>
 
 <h2>Places</h2>
-<p>You can either <a href="/admin/places/new">create</a> new places, or <a href="/admin/places/edit">edit</a> them.</p>
+<p>You can either <a href="/admin/places/create">create</a> new places, or <a href="/admin/places/">edit</a> them.</p>
 @stop