+++ changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Version 0.3.1 (2017-03-03)
+   - Correct command to restart daemon queues in deploy.sh
+   - Improve Admin CP by “resource”-ifying the controllers
+
 ## Version 0.3 (2017-03-02)
   - convert env() calls to config() calls for cacheing
   - refactor routes and give important one names